// NoelApp/src/model/mode.ts
export type RootStackParamList = {
    Home: undefined;
    Details: undefined;
    CreateCard: undefined;
    CountdownScreen: undefined;
    CardScreen: undefined;
    SelfieScreen: undefined;
    PhotoLibraryScreen: undefined;
    SettingScreen: undefined;
    Sticker:undefined;
    Color:undefined;
    Image:undefined;
    Font:undefined;
    ScreenCard: { image: any }; 
    
  };
  